+++ CHANGELOG.md
@@ -1,8 +1,27 @@
 # Changelog
 
+## Astro version 1.1.0
+
+This is the changelog for Astro version 1.1.0 released on ____, 2024.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/astro/tags)
+
+### Breaking Change
+
+* `:tz_world` is no longer a required dependency - it is now an optional dependency. This library is used to resolve a time zone name from a given latitude and longitude. When configured, it becomes the default method of resolving time zone names from a location. However it is no possible to provide alternative implementations for this resolution using the `:time_zone_resolver` option.
+
+* To retain the previous behaviour, applications should add `{:tz_world, "~> 1.0"}` to their dependencies.
+
+### Enhancements
+
+* Adds an option `:time_zone_resolver` to `Astro.sunrise/3` and `Astro.sunset/3` that is invoked to resolve the time zone name from a given latitude and longitude. The default is to use `TzWorld.timezone_at/1` if `TzWorld` is configured, otherwise an error is returned.
+
+* The default time zone database is now detected in the following order:
+    Application.get_env(:elixir, :time_zone_database) 
+    TzData.TimeZoneDatabase if TzData is configured
+    Tz.TimeZoneDatabase if Tz is configured
+
 ## Astro version 1.0.2
 
-This is the changelog for Astro version 1.0.2 released on ay 27th, 2024.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/astro/tags)
+This is the changelog for Astro version 1.0.2 released on May 27th, 2024.  For older changelogs please consult the release tag on [GitHub](https://github.com/kipcole9/astro/tags)
 
 ### Bug Fixes